requirements.txt
cffi==1.15.1
cryptography==41.0.3
dnspython==2.4.2
pycparser==2.21
pyOpenSSL==23.2.0
